refactor: CreateChallenge 컴포넌트 폴더명을 kebab-case로 변경 #27
