+++ src/components/create-challenge/MaxParticipantsSheet.tsx
