+++ src/components/create-challenge/VerificationDaysSheet.tsx
